♻️ Update push auto config to be static
